+++ CHANGELOG.md
@@ -13,7 +13,7 @@
   - Feat: add `ActionInput.peek`.
   - Feat: add `ActionBuilder` for chaining action decorators.
   - Feat: add `Builder.useError` to set error type.
-  - Feat: add `Builder.branch` to allow an action to map multiple token kinds. [#28](https://github.com/DiscreteTom/retsac/issues/28)
+  - Feat: add `Builder.select` to allow an action to map multiple token kinds. [#28](https://github.com/DiscreteTom/retsac/issues/28)
     - Add `Action.kinds`, `ActionWithKinds` and `SelectedAction`.
   - Feat: add `Lexer.javascript`.
     - Add `evalString`. [#30](https://github.com/DiscreteTom/retsac/issues/30)